Organizing TUSClient tests
