+++ Tests/TUSKitTests/Mocks.swift
@@ -16,6 +16,8 @@ final class TUSMockDelegate: TUSClientDelegate {
     var failedUploads = [(UUID, Error)]()
     var fileErrors = [TUSClientError]()
     
+    var activityCount = { finishedUploads.count + startedUploads.count + failedUploads.count + fileErrors.count }
+    
     var finishUploadExpectation: XCTestExpectation?
     var startUploadExpectation: XCTestExpectation?
     var fileErrorExpectation: XCTestExpectation?
